requirements.txt
fastapi
uvicorn
streamlit
chromadb
sentence-transformers
langchain
langchain-community
langchain-huggingface
langchain-chroma
groq
python-dotenv
pypdf